Moving LoRaAPIHelper to common
