+++ Tests/Common/LoRaAPIHelper.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace LoRaWan.Tests.Simulation
+namespace LoRaWan.Tests.Common
 {
     using System;
     using System.Net.Http;
